backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
click==8.1.8
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.12
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
openai==1.68.2
psycopg2-binary==2.9.9
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.1.0
sniffio==1.3.1
starlette==0.46.1
tqdm==4.67.1
typing_extensions==4.13.0
uvicorn==0.34.0
